LayoutTests/platform/chromium-mac/http/tests/websocket/tests/hybi/bufferedAmount-after-close-expected.txt

Web Socket bufferedAmount after closed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
Closed.
PASS ws.readyState is 3
PASS ws.bufferedAmount is 0
PASS ws.send('send to closed socket') is false
FAIL ws.bufferedAmount should be 23. Was 27.
PASS successfullyParsed is true

TEST COMPLETE
